requirements.txt
# api/requirements.txt
Flask
gunicorn
requests
python-dotenv